requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anthropic==0.75.0
    # via
    #   prompt2mesh (pyproject.toml)
    #   langchain-anthropic
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
bcrypt==5.0.0
    # via prompt2mesh (pyproject.toml)
blinker==1.9.0
    # via streamlit
cachetools==6.2.2
    # via streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   typer
    #   uvicorn
cryptography==46.0.3
    # via pyjwt
distro==1.9.0
    # via anthropic
docker==7.1.0
    # via prompt2mesh (pyproject.toml)
docstring-parser==0.17.0
    # via anthropic
fastapi==0.122.0
    # via prompt2mesh (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   langgraph-sdk
    #   langsmith
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.12.0
    # via anthropic
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via
    #   altair
    #   mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==1.1.0
    # via prompt2mesh (pyproject.toml)
langchain-anthropic==1.2.0
    # via prompt2mesh (pyproject.toml)
langchain-core==1.1.0
    # via
    #   prompt2mesh (pyproject.toml)
    #   langchain
    #   langchain-anthropic
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==1.0.4
    # via
    #   prompt2mesh (pyproject.toml)
    #   langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.2.10
    # via langgraph
langsmith==0.4.49
    # via
    #   prompt2mesh (pyproject.toml)
    #   langchain-core
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mcp==1.22.0
    # via prompt2mesh (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
narwhals==2.12.0
    # via altair
numpy==2.3.5
    # via
    #   pandas
    #   pydeck
    #   streamlit
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.0.0
    # via streamlit
protobuf==6.33.1
    # via streamlit
psycopg2-binary==2.9.11
    # via prompt2mesh (pyproject.toml)
pyarrow==21.0.0
    # via streamlit
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   anthropic
    #   fastapi
    #   langchain
    #   langchain-anthropic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via
    #   prompt2mesh (pyproject.toml)
    #   mcp
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   prompt2mesh (pyproject.toml)
    #   mcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   prompt2mesh (pyproject.toml)
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via typer
rpds-py==0.29.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sqlalchemy==2.0.44
    # via prompt2mesh (pyproject.toml)
sse-starlette==3.0.3
    # via mcp
starlette==0.50.0
    # via
    #   fastapi
    #   mcp
streamlit==1.51.0
    # via prompt2mesh (pyproject.toml)
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typer==0.20.0
    # via mcp
typing-extensions==4.15.0
    # via
    #   altair
    #   anthropic
    #   anyio
    #   fastapi
    #   langchain-core
    #   mcp
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   streamlit
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via mcp
watchdog==6.0.0
    # via streamlit
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
